Add purls flag to fix command
