+++ src/commands/fix/shared.ts
@@ -1,6 +1,4 @@
-import semver from 'semver'
-
-import type { FixOptions, NormalizedFixOptions, RangeStyle } from './types'
+import type { FixOptions, NormalizedFixOptions } from './types'
 
 export const CMD_NAME = 'socket fix'
 
@@ -27,36 +25,3 @@ export function assignDefaultFixOptions(
   }
   return options as NormalizedFixOptions
 }
-
-export function applyRange(
-  refRange: string,
-  version: string,
-  style: RangeStyle = 'preserve'
-): string {
-  switch (style) {
-    case 'caret':
-      return `^${version}`
-    case 'gt':
-      return `>${version}`
-    case 'gte':
-      return `>=${version}`
-    case 'lt':
-      return `<${version}`
-    case 'lte':
-      return `<=${version}`
-    case 'preserve': {
-      const comparators = [...new semver.Range(refRange).set].flat()
-      const { length } = comparators
-      return !length || length > 1
-        ? version
-        : `${comparators[0]!.operator}${version}`
-    }
-    case 'tilde':
-      return `~${version}`
-    case 'pin':
-    default:
-      return version
-  }
-}
-
-export const RangeStyles = ['caret', 'gt', 'lt', 'pin', 'preserve', 'tilde']
