+++ src/commands/fix/types.ts
@@ -1,23 +1,15 @@
+import type { RangeStyle } from '../../utils/semver'
 import type { Spinner } from '@socketsecurity/registry/lib/spinner'
 
 type StripUndefined<T> = {
   [K in keyof T]-?: Exclude<T[K], undefined>
 }
 
-export type RangeStyle =
-  | 'caret'
-  | 'gt'
-  | 'gte'
-  | 'lt'
-  | 'lte'
-  | 'pin'
-  | 'preserve'
-  | 'tilde'
-
 export type FixOptions = {
   autoMerge?: boolean | undefined
   autoPilot?: boolean | undefined
   cwd?: string | undefined
+  purls?: string[] | readonly string[] | undefined
   rangeStyle?: RangeStyle | undefined
   spinner?: Spinner | undefined
   test?: boolean | undefined
